made AddToCart button work

+++ spring/src/main/java/com/ibooku/kickoff/service/BookRepository.java
@@ -7,6 +7,6 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import com.ibooku.kickoff.model.Book;
 
 @CrossOrigin(origins="http://localhost:4200")
-public interface BookRepository extends CrudRepository<Book, Long> {
+public interface BookRepository extends CrudRepository<Book, Integer> {
 	Book findByBId(Integer id);
 }
